+++ Prototyping/Map Tile Images/Prototype - Map Tile Images.cpp
@@ -1,4 +1,4 @@
-#include <Simp1e/Qt.h>
+#include <Simp1e/UI/Qt/Maps.h>
 
 using namespace Simp1e::Maps;
 using namespace Simp1e::UI;
