+++ Simp1e.Qt.Maps/include/Simp1e/UI/Qt/QtTile.h
