--- Simp1e.Qt/include/Simp1e/Qt.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "UI/Qt.h"
